Various improvements to ExpectationNode.hh header based on PR comments
